# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CEX Crypto Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      parameters:
        - name: pair
          in: query
          schema:
            type: string
          description: Trading pair (e.g. BTC/USDT)
        - name: status
          in: query
          schema:
            type: string
            enum: [open, filled, cancelled, partial]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: User orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    delete:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        pair:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [limit, market]
        status:
          type: string
          enum: [open, filled, cancelled, partial]
        price:
          type: string
          nullable: true
        quantity:
          type: string
        filled:
          type: string
        total:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - pair
        - side
        - type
        - status
        - quantity
        - filled
        - total
        - createdAt
        - updatedAt
    CreateOrderBody:
      type: object
      properties:
        pair:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [limit, market]
        price:
          type: string
          nullable: true
        quantity:
          type: string
      required:
        - pair
        - side
        - type
        - quantity
...